refactor: move merge function back into taskfile package
